Conflicts:
	src/main/java/ba/etf/tim11/eCinema/dao/CommentDao.java
	src/main/java/ba/etf/tim11/eCinema/dao/DaoFactory.java
	src/main/java/ba/etf/tim11/eCinema/dao/impl/ContentDaoImpl.java

+++ src/main/resources/ba/etf/tim11/eCinema/resources/GroupResource.java
@@ -16,6 +16,10 @@ import ba.etf.tim11.eCinema.dao.DaoFactory;
 import ba.etf.tim11.eCinema.dao.GroupDao;
 import ba.etf.tim11.eCinema.dao.impl.JDBCDaoFactory;
 import ba.etf.tim11.eCinema.models.Group;
+import ba.etf.tim11.eCinema.resources.responses.BadRequestException;
+import ba.etf.tim11.eCinema.resources.responses.ResourceNotFoundException;
+import ba.etf.tim11.eCinema.resources.responses.Response;
+import ba.etf.tim11.eCinema.utils.ResourceUtil;
 
 
 @Path("groups")
@@ -43,27 +47,66 @@ public class GroupResource
 	@Path("{id}")
 	public Group getGroup(@PathParam("id") int id) 
 	{
-		return groupDao.find(id);
+		Group group = groupDao.find(id);
+		
+		if (group == null) {
+			throw new ResourceNotFoundException("Group not found");
+		}
+		
+		return group;
 	}
 	
 	@POST
 	@Consumes("application/x-www-form-urlencoded")
 	public Group createNewGroup(MultivaluedMap<String, String> formParams) 
 	{
+		if (!ResourceUtil.hasAll(formParams, "name", "description")) {
+			throw new BadRequestException("You are missing name and/or description field.");
+		}
+		
 		Group group = new Group();
 		
 		group.setName(formParams.getFirst("name"));
 		group.setDescription(formParams.getFirst("description"));
 
-		return groupDao.insert(group);
+		groupDao.insert(group);
+		
+		return group;
+	}
+	
+	@POST
+	@Path("{id}")
+	@Consumes("application/x-www-form-urlencoded")
+	public Response updateGroup(@PathParam("id") int id, MultivaluedMap<String, String> formParams) 
+	{	
+		Group group = groupDao.find(id);
+		if (group == null) {
+			throw new ResourceNotFoundException("Group not found.");
+		}
+		
+		if (formParams.getFirst("name") != null)
+			group.setName(formParams.getFirst("name"));
+		
+		if (formParams.getFirst("description") != null)
+			group.setDescription(formParams.getFirst("description"));
+
+		groupDao.update(group);
+		
+		return Response.success();
 	}
 	
 	@DELETE
 	@Path("{id}")
-	public void deleteUser(@PathParam("id") int id) 
+	public Response deleteUser(@PathParam("id") int id) 
 	{
 		Group group = groupDao.find(id);
+		if (group == null) {
+			throw new ResourceNotFoundException("User not found.");
+		}
+		
 		groupDao.delete(group);
+		
+		return Response.success();
 	}
 
 }
